src/binanalyzer_database/requirements.txt
# pandas
flask==3.0.2
gunicorn==21.2.0
pydantic
# mysql==0.0.3
mysql-connector-python==8.4.0
PyMySQL==1.1.1
sqlmodel==0.0.19